requirements.txt
django>=4.2,<5.0
djangorestframework>=3.14
pdfplumber>=0.10
faiss-cpu>=1.7
google-generativeai>=0.5
numpy>=1.24
python-dotenv>=1.0
